requirements.txt
fastapi
mongoengine
scikit-learn
uvicorn
python-dotenv
pandas
redis
pytest
apscheduler